+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 next
 ----
 
+- Mark Matplotlib 3.10.3 as incompatible.
 - Require registration in ``mpl.rcParams["figure.hooks"]`` for the
   :envvar:`MPLCURSORS` environment variable.
 - Support new-style (Matplotlib≥3.8) contour plots.